Revert "static analyser v0.3"
